Remodel nfa with Option alphabet
Instead of duplicating logic for non-deterministic graphs with epsilon
edges and without, model the nfa underlying structure agnostic of the
alphabet extension above. eNFAs are then non-deterministic graphs over
the alphabet Option<A>. This is an implementation detail, otherwise we
would instead expose a new type `Epsilon<A>` instead of using the
standard Option type.